+++ src/tool/CHANGELOG.md
@@ -2,15 +2,16 @@
 
 (This log only contains changes for grebedb-tool. Changes for the library are located in its own file.)
 
-## Unreleased
+## 0.3.0 (2021-04-20)
 
 * Added `--zstd` flag to `import` and `export` commands.
 * Added `inspect` command.
+* Updated dependency to grebedb 0.3
 
 ## 0.2.0 (2021-04-07)
 
-* Updated dependency to grebedb 0.2.0
 * Added `verify` command.
+* Updated dependency to grebedb 0.2
 
 ## 0.1.0 (2021-03-28)